Hack to spawn play server from Go CI.
Go, actually java, waits for the process to quit to return its
status, as result if you try to start "play" direct from Go it
will not stop building. This bash script spawns play server on a
subshell and then waits to see if the server went up signaling
Go properly.